requirements.txt
adodbapi==2.6.0.7
backcall==0.1.0
colorama==0.3.9
decorator==4.3.0
Django==2.1.1
django-bootstrap-form==3.4
django-bootstrap3==11.0.0
django-mssql==1.8
django-pymssql==1.7.1
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
djangorestframework==3.8.2
ipdb==0.11
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
numpy==1.15.4
pandas==0.23.4
parso==0.3.1
pickleshare==0.7.5
prompt-toolkit==2.0.5
Pygments==2.2.0
pymssql==2.1.4
pyodbc==4.0.24
pypiwin32==223
python-dateutil==2.7.5
pytz==2018.5
pywin32==223
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
